create error_messages and access restrictions

+++ app/Http/Controllers/TasksController.php
@@ -15,7 +15,8 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        // $tasks = Task::all();
+        $tasks = Task::orderBy('id', 'desc')->paginate(10);
         return view('tasks.index', ['tasks' => $tasks]);
     }
 
@@ -46,6 +47,7 @@ class TasksController extends Controller
         $task = new Task;
         $task->content = $request->content;
         $task->status = $request->status;
+        $task->user_id = $request->user()->id;
         $task->save();
         
         return redirect('/');
@@ -60,8 +62,16 @@ class TasksController extends Controller
      */
     public function show($id)
     {
+        
         $task = Task::findOrFail($id);
-        return view('tasks.show', ['task' => $task]);
+        
+        if ($task->user_id == \Auth::id()){
+            return view('tasks.show', ['task' => $task]);
+        }else{
+            return redirect('/');
+        }
+        
+        
     }
 
     /**
@@ -73,7 +83,13 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::findOrFail($id);
-        return view('tasks.edit', ['task' => $task]);
+        
+        if ($task->user_id == \Auth::id()){
+            return view('tasks.edit', ['task' => $task]);
+        }else{
+            return redirect('/');
+        }
+            
     }
 
     /**
@@ -91,10 +107,16 @@ class TasksController extends Controller
         ]);
 
         $task = Task::findOrFail($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        return redirect('/');
+        
+        if ($task->user_id == \Auth::id()){
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+            return redirect('/');
+        }else{
+            return redirect('/');
+        }
+        
     }
 
     /**
@@ -106,7 +128,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
-        return redirect('/');
+        
+        if($task->user_id == \Auth::id()){
+            $task->delete();
+            return redirect('/');
+        }else{
+            return redirect('/');
+        }
+        
     }
 }
